imgmarker/pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "imgmarker"
version = "0.1.0"
readme = "README.md"

dependencies = [
  "astropy==6.1.3",
  "pillow==10.4.0",
  "PyQt6==6.7.1",
]

authors = [
  {name = "Andi Kisare", email = "akisare@outlook.com"},
  {name = "Ryan Walker", email = "astrorya@umich.edu"}
]

[tool.setuptools.packages.find]
where = [".."]
include = ["imgmarker"]

[tool.setuptools.package-data]
imgmarker = ["*.png"]

[project.scripts]
imgmarker = "imgmarker.__init__:main"


